Adjust syscalls to thread-safe ones. Start with stackalloc, then use loop.

+++ src/native/libs/System.Native/pal_uid.h
@@ -85,14 +85,14 @@ PALEXPORT int32_t SystemNative_GetGroups(int32_t ngroups, uint32_t* groups);
 
 /**
 * Gets the user name associated with the specified user ID and stores it in the buffer.
-* On failure, returns -1 and sets the buffer to NULL.
-* On success, returns the length of the uname string, and sets buffer to a valid pointer.
+* On failure, returns -1 and sets errno.
+* On success, return the length of the buffer, and the buffer contains the uname.
 */
-PALEXPORT int32_t SystemNative_GetUName(uint32_t uid, char* buffer, int32_t bufferSize);
+PALEXPORT int32_t SystemNative_GetUName(uint32_t uid, char* buffer, size_t bufferSize);
 
 /**
 * Gets the user name associated with the specified group ID and stores it in the buffer.
-* On failure, returns -1 and sets the buffer to NULL.
-* On success, returns the length of the gname string, and sets buffer to a valid pointer.
+* On failure, returns -1 and sets errno.
+* On success, return the length of the buffer, and the buffer contains the gname.
 */
-PALEXPORT int32_t SystemNative_GetGName(uint32_t gid, char* buffer, int32_t bufferSize);
+PALEXPORT int32_t SystemNative_GetGName(uint32_t gid, char* buffer, size_t bufferSize);
